requirements.txt
anyio==4.4.0
APScheduler==3.6.3
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==4.2.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
h11==0.14.0
html5lib==1.0.1
httpcore==1.0.5
httplib2==0.9.2
httpx==0.27.2
idna==3.8
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tornado==6.1
tzlocal==5.2
urllib3==2.2.2
webencodings==0.5.1
